feat: 유효성 검사 및 예외처리

+++ src/main/java/com/study/profile_stack_api/domain/profile/service/ProfileService.java
@@ -3,8 +3,11 @@ package com.study.profile_stack_api.domain.profile.service;
 import com.study.profile_stack_api.domain.profile.dao.ProfileDao;
 import com.study.profile_stack_api.domain.profile.dto.ProfileRequest;
 import com.study.profile_stack_api.domain.profile.dto.ProfileResponse;
+import com.study.profile_stack_api.domain.profile.entity.Position;
 import com.study.profile_stack_api.domain.profile.entity.Profile;
 import com.study.profile_stack_api.global.common.PageResponse;
+import com.study.profile_stack_api.global.exception.DataNotFoundException;
+import com.study.profile_stack_api.global.exception.DuplicateEmailException;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -18,11 +21,18 @@ public class ProfileService {
     private final ProfileDao profileDao;
 
     public ProfileResponse createProfile(ProfileRequest profileRequest) {
+
+        Position position = Position.of(profileRequest.getPosition());
+        String email = profileRequest.getEmail();
+
+        if(profileDao.existsByEmail(email))
+            throw new DuplicateEmailException(email);
+
         Profile profile = Profile.builder()
                 .name(profileRequest.getName())
-                .email(profileRequest.getEmail())
+                .email(email)
                 .bio(profileRequest.getBio())
-                .position(profileRequest.getPosition())
+                .position(position.name())
                 .careerYears(profileRequest.getCareerYears())
                 .githubUrl(profileRequest.getGithubUrl())
                 .blogUrl(profileRequest.getBlogUrl())
@@ -49,19 +59,27 @@ public class ProfileService {
     }
 
     public ProfileResponse getProfile(Long id) {
-        Profile profile = profileDao.findById(id);
+        Profile profile = profileDao.findById(id)
+                .orElseThrow(() -> new DataNotFoundException("프로필", id));
 
         return ProfileResponse.from(profile);
     }
 
     @Transactional
     public ProfileResponse updateProfile(Long id,ProfileRequest profileRequest) {
-        Profile profile = profileDao.findById(id);
+        Position position = Position.of(profileRequest.getPosition());
+        String email = profileRequest.getEmail();
+
+        if(profileDao.existsByEmail(email))
+            throw new DuplicateEmailException(email);
+
+        Profile profile = profileDao.findById(id)
+                .orElseThrow(() -> new DataNotFoundException("프로필", id));
 
         profile.setName(profileRequest.getName());
         profile.setEmail(profileRequest.getEmail());
         profile.setBio(profileRequest.getBio());
-        profile.setPosition(profileRequest.getPosition());
+        profile.setPosition(position.getDescription());
         profile.setCareerYears(profileRequest.getCareerYears());
         profile.setGithubUrl(profileRequest.getGithubUrl());
         profile.setBlogUrl(profileRequest.getBlogUrl());
@@ -72,12 +90,15 @@ public class ProfileService {
     public List<ProfileResponse> findByPosition(String position) {
         List<Profile> profiles = profileDao.findByPosition(position);
 
-        return profiles.stream()
-                .map(ProfileResponse::from)
-                .toList();
+        return (profiles != null) ?
+                profiles.stream().map(ProfileResponse::from).toList()
+                : List.of();
     }
 
     public void deleteProfile(Long id) {
+        if(!profileDao.existsById(id))
+            throw new DataNotFoundException("프로필", id);
+
         profileDao.deleteProfile(id);
     }
 }
